+++ WhirlyGlobeSrc/WhirlyGlobeLib/include/MaplyPanDelegate.h
@@ -28,11 +28,11 @@
     /// Set if we're panning
     BOOL panning;
     /// View transform when we started
-    Eigen::Matrix4f startTransform;
+    Eigen::Matrix4d startTransform;
     /// Where we first touched the plane
-    WhirlyKit::Point3f startOnPlane;
+    WhirlyKit::Point3d startOnPlane;
     /// Viewer location when we started panning
-    WhirlyKit::Point3f startLoc;
+    WhirlyKit::Point3d startLoc;
     CGPoint lastTouch;
     /// Boundary quad that we're to stay within
     std::vector<WhirlyKit::Point2f> bounds;
